Allow individual changes to the x, y, z rotation values.
Instead of spinning use left/right arrows keys to individually change
the rotation on an axis by +1/-1 degree. With right-ctrl select Z,
left-ctrl selects X and left-shift selects Y.

It seems a "left hand rule" is being used as wall as the axis is
stationary and the monkey moves. In my 3D-soft-engine the axis and the
monkey both move and is "right hand rule" for Y and Z but "left hand rule"
of X. And X being different is definetly WRONG!